requirements.txt
streamlit>=1.34
requests>=2.31
pillow>=10.0
